Created a Component for Categories Dropdown
