+++ src/pages/product/searchProducts/SearchFilters.tsx
@@ -1,20 +1,10 @@
-import { useEffect, useMemo, useRef, useState } from "react";
+import { useEffect, useRef, useState } from "react";
 import Checkbox from "../../../components/input/Checkbox";
 import useQueryParams from "../../../hooks/useQueryParams";
-import {
-  CheckedIcon,
-  DoubleLayerIcon,
-  PercentIcon,
-  RupeesIcon,
-  SingleLayerIcon,
-  TripleLayerIcon,
-} from "../../../icons";
-import { CATEGORIES_DATA } from "../../../constants/categories";
-import { DEFAULT_FILTER } from "../../../constants";
+import { CheckedIcon, PercentIcon, RupeesIcon } from "../../../icons";
 import Dropdown from "../../../components/dropdown/Dropdown";
-import { TDropdownOption } from "../../../types";
-import DropdownOptions from "./components/DropdownOptions";
 import Input from "../../../components/input/Input";
+import DropdownCategories from "./components/DropdownCategories";
 
 interface FiltersProps {
   className?: string;
@@ -157,112 +147,9 @@ function SearchFilters({ className = "" }: FiltersProps) {
     // eslint-disable-next-line react-hooks/exhaustive-deps
   }, [ranges.discount]);
 
-  const selectedCategories = useMemo(
-    () => ({
-      category_1: queryParams.category_1,
-      category_2: queryParams.category_2,
-      category_3: queryParams.category_3,
-    }),
-    [queryParams]
-  );
-
-  const handleFilterChange = useMemo(
-    () => ({
-      category_1: (val: string) => {
-        if (val === queryParams.category_1 || val === DEFAULT_FILTER.value) {
-          removeParam("category_1");
-        } else {
-          setParams({ category_1: val });
-        }
-        removeParam("category_2");
-        removeParam("category_3");
-      },
-      category_2: (val: string) => {
-        if (val === queryParams.category_2 || val === DEFAULT_FILTER.value) {
-          removeParam("category_2");
-        } else {
-          setParams({ category_2: val });
-        }
-        removeParam("category_3");
-      },
-      category_3: (val: string) => {
-        if (val === queryParams.category_3 || val === DEFAULT_FILTER.value) {
-          removeParam("category_3");
-        } else {
-          setParams({ category_3: val });
-        }
-      },
-    }),
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-    [queryParams.category_1, queryParams.category_2, queryParams.category_3]
-  );
-
-  const { level1Options, level2Options, level3Options } = useMemo(() => {
-    const level1Options = CATEGORIES_DATA || [];
-    const level1Data = level1Options.find(
-      (cat) => cat.category === selectedCategories.category_1
-    );
-
-    const level2Options = level1Data?.subCategories || [];
-
-    const level2Data = level1Data?.subCategories.find(
-      (cat) => cat.category === selectedCategories.category_2
-    );
-
-    const level3Options = level2Data?.subCategories || [];
-
-    return { level1Options, level2Options, level3Options };
-  }, [selectedCategories]);
-
-  const CATEGORY_OPTIONS = useMemo(
-    () => [
-      {
-        onChange: (category: TDropdownOption) =>
-          handleFilterChange.category_1(category.value),
-        heading: {
-          title: "Category One",
-          icon: SingleLayerIcon,
-        },
-        selected: selectedCategories.category_1,
-        options: level1Options.map((opt) => ({
-          name: opt.name,
-          value: opt.category,
-        })),
-      },
-      {
-        onChange: (category: TDropdownOption) =>
-          handleFilterChange.category_2(category.value),
-        heading: {
-          title: "Category Two",
-          icon: DoubleLayerIcon,
-        },
-        selected: selectedCategories.category_2,
-        options: level2Options.map((opt) => ({
-          name: opt.name,
-          value: opt.category,
-        })),
-      },
-      {
-        onChange: (category: TDropdownOption) =>
-          handleFilterChange.category_3(category.value),
-        heading: {
-          title: "Category Three",
-          icon: TripleLayerIcon,
-        },
-        selected: selectedCategories.category_3,
-        options: level3Options.map((opt) => ({
-          name: opt.name,
-          value: opt.category,
-        })),
-      },
-    ],
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-    [handleFilterChange]
-  );
-
   return (
     <section
-      className={`h-full gap-6 bg-primary-inverted select-none ${className}`}
+      className={`h-full bg-primary-inverted select-none ${className}`}
     >
       <div className={`w-full flex flex-col gap-1 py-2 px-6`}>
         {/* Availability Filter */}
@@ -430,30 +317,7 @@ function SearchFilters({ className = "" }: FiltersProps) {
             </div>
           </>
         </Dropdown>
-        {CATEGORY_OPTIONS.map((cat, index) => {
-          const HeadingIcon = cat.heading.icon;
-          return (
-            <Dropdown
-              key={index}
-              heading={{
-                title: cat.heading.title,
-                icon: (
-                  <>
-                    (
-                    <HeadingIcon className="w-4 h-4 -m-[1.5px] stroke-primary" />
-                    )
-                  </>
-                ),
-              }}
-            >
-              <DropdownOptions
-                onChange={(data) => cat.onChange(data)}
-                selected={cat.selected}
-                options={[DEFAULT_FILTER, ...cat.options]}
-              />
-            </Dropdown>
-          );
-        })}
+        <DropdownCategories />
       </div>
     </section>
   );
